Update Festivals of india
The project aims to explore the rich cultural heritage of India by creating a computer 
graphics and visualization project focused on its vibrant festivals. The project will utilize 
various computer graphics techniques and visualization tools to showcase the unique and 
diverse celebrations that take place across the country, offering an immersive and interactive 
experience for viewers.